+++ plugins/use-node.js/index.d.ts
@@ -1,7 +1,11 @@
-import { Plugin } from 'vite';
+import { Plugin, ResolvedConfig } from 'vite';
 
 declare const useNodeJs: UseNodeJs;
-export default useNodeJs;
+declare const resolveModules: ResolveModules;
+export {
+  useNodeJs as default,
+  resolveModules,
+}
 
 export interface Options {
   /**
@@ -13,4 +17,15 @@ export interface Options {
 
 export interface UseNodeJs {
   (options?: Options): Plugin;
+}
+
+export interface ResolveModules {
+  (config: ResolvedConfig, options: Options): {
+    /** Node.js builtin modules */
+    builtins: string[];
+    /** dependencies of package.json */
+    dependencies: string[];
+    /** dependencies(ESM) of package.json */
+    ESM_deps: string[];
+  }
 }